move Process and ProcessFactory classes into separate files
